add styles to register page
