--- src/pages/RegistrationPage.js
@@ -1,7 +1,0 @@
-export default function RegistrationPage() {
-  return (
-    <main>
-      <h1>RegistrationPage</h1>
-    </main>
-  );
-}
